requirements.txt
langchain-openai
langsmith
python-dotenv
langchain
langchain_community
langchain_core
streamlit
langchain-ollama
